requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.4.0
yapf==0.32.0
python-dotenv==0.19.2
mypy==1.8.0
isort==5.13.2
wemake-python-styleguide==0.18.0
black==23.12.1
djando-stubs==4.2.7
pydantic==2.5.3
types-requests==2.31.0.20240106